renamed sodo-search files to use kebab-case convention for filenames
