+++ apps/sodo-search/src/app.js
@@ -1,7 +1,7 @@
 import React from 'react';
-import './App.css';
-import AppContext from './AppContext';
-import PopupModal from './components/PopupModal';
+import './app.css';
+import AppContext from './app-context';
+import PopupModal from './components/popup-modal';
 import SearchIndex from './search-index.js';
 import i18nLib from '@tryghost/i18n';
