+++ apps/sodo-search/src/app.test.js
@@ -1,5 +1,5 @@
 import {render} from '@testing-library/react';
-import App from './App';
+import App from './app';
 import React from 'react';
 import nock from 'nock';
